SpitNPolish/Bug Fixes 1
Changes:
Boss, Chihuahua, Husky, Secuirty, and Tanuki.tscn, Enemy3.cd, NPC.cs: I noticed that the Player could not tell what HP a NPC was at. As such I added a Progress Bar above the enemy NPC heads to show this to the player. The tanuki one does not show if transformed.
SecuirtyBoss.tscn - The tanuki in this room was obstructing the player from entering the boss room so I moved him up a bit.
Settings.tscn, SettingsMenu.cs - I noticed that when the volume bars we set to there lowest point the volume would not mute. For now I have extended the range of the bars to have -60db as the minimum (which is effectivly muted). This has other issues but works for now.
Player.cs - In the most recent version there was an error that still identified the WinLoseMenu as a ColorRect instead of a Panel. This was causing the game to freak out on player win/lose. I simply changed it.