Bug Fixes
Found a bug in the boss script. It was making the explosions a child of the root instead of the scene, resulting in them carrying over between scenes and messing up player spawning. At least player death no longer causes problems. I also fixed that box on the docks since every single play tester had to deal with it and it would only take 5 seconds to fix. I did find a bug with the shotgun that made it so that it didn't get equiped or picked up for some reason. I'm going to fix that, but in the mean time, I am going to push.