Added attack radius to melee npcs, directional bullet movement, and changed the difficulty values
I have done the following
I changed the difficulty values to .4, .3, and .2. These felt fine when I was playing but we can scale them back easily if needed.
There was a small bug with the Dock scene for scene transition where its fade out was not working. I believe I fixed it.
I added some test code for the bullets moving to the players mouse. The original code is still there, but it is currently set up to work with what I had written. This new CreateProjectileTEST function gets the mouses current postion and direction relative to the player. It then uses this to determine what direction the bullet needs to be facing and moving.
I added a visual aid for the melee attacks for Paul and the Chihuahuas. They both now have a color rect that moves with their attack range (transparent when not in combat). When in combat, and the player is in range, it will start to fill. Once filled then they attack.