Fixes to Sus calculation, and additional ways to generate sus
Changes:
Dashing, and any player attack now increases the local room's sus by 2%.
Local Thresholds were not being communicated to the player on the UI, so I added a color rect to it. The local sus progress bar now has a color rect overlayed on top of it. It is positioned at its appropriate point (and also updates it position if Paul moves it). I made the color a transparent blue but we can change it later.
3a. After implementing the color rect I realized 2 things. One there was an error in the IncreaseLocalSuspicion function. This error was adding to GlobalSus even if over the max. This was not a problem before because of how we were adding to Global, but with the addition of the weapons and dashing it messed it up. As such I rearranged the order of the function so it only performs the calculations/addition if it is supposed to.
3b. Two, MaxGlobalSuspicion was not being calculated correctly. From my understanding, MaxGlobalSuspicion was supposed to be the sum of all rooms maxSus - there thresholds (as that was how IncreaseLocalSuspicion functioned). However it was instead just a sum of all room suspicions. To fix this I just changed the calculation in Game.cs to reflect what it should be.