Adjusted UI opacity and Difficulty Values
Changes:
Changed the opacity of the grey portion of the UI, giving the player more visibility of the room.
In relation to the difficulty values on the hard difficulty I almost lost to Global Sus during several runs, and on Easy I would get to the boss with around 20 to 50% to spare. This is from a person who knows what they are doing though so I scaled the difficulties back by 5% each (Easy: 0.45, Medium 0.35, Hard: 0.25).